Codegen for long constants
git-svn-id: https://svn.uni-koblenz.de/gupro/re-group/trunk/project/grabaja@7881 6c03965d-96ef-0310-af27-ce2fa9dd3b5a

+++ testit/cginput/bar/SampleAnnotation.java
@@ -2,7 +2,7 @@ package bar;
 
 @SampleAnnotation
 public @interface SampleAnnotation {
-	public int number() default -1;
+	public long number() default -1l;
 
 	public String text() default "";
